+++ scripts/delete-chunk.cjs
@@ -2,16 +2,22 @@ const fs = require("node:fs");
 
 let chunkName = process.argv[3];
 
-fs.unlinkSync(`./src/chunks/${chunkName}.tsx`);
+const deleteFile = (path) => {
+  if (fs.existsSync(path)) {
+    fs.unlinkSync(path);
+  }
+};
 
-fs.unlinkSync(`./src/css/${chunkName}.css`);
+deleteFile(`./src/chunks/${chunkName}.tsx`);
+deleteFile(`./src/css/${chunkName}.css`);
+deleteFile(`./src/html/${chunkName}.html`);
 
-fs.unlinkSync(`./src/html/${chunkName}.html`);
-
-const data = fs.readFileSync("./templateconfig.json");
-const json = JSON.parse(data);
-const basenames = json.basenames;
-if (basenames.indexOf(chunkName) !== -1) {
-  json.basenames.splice(basenames.indexOf(chunkName), 1);
-  fs.writeFileSync("./templateconfig.json", JSON.stringify(json));
+if (fs.existsSync("./templateconfig.json")) {
+  const data = fs.readFileSync("./templateconfig.json");
+  const json = JSON.parse(data);
+  const basenames = json.basenames;
+  if (basenames.indexOf(chunkName) !== -1) {
+    json.basenames.splice(basenames.indexOf(chunkName), 1);
+    fs.writeFileSync("./templateconfig.json", JSON.stringify(json));
+  }
 }
